@@ -1,21 +1,25 @@
+import { Box, Button, Paper } from '@material-ui/core'
 import React, { Component } from 'react'
+import NumberFormat from 'react-number-format'
 import { Product } from '../../model/Product'
 import { cartService } from '../../service/CartService'
+import SearchIcon from '@material-ui/icons/Search';
 
 export default class ProductItem extends Component<Props, State> {
     render() {
-        const pcl = Number(this.props.product.salePrice)
-        const pc = Number(this.props.product.price)
+        // const pcl = Number(this.props.product.salePrice)
+        // const pc = Number(this.props.product.price)
 
-        const priceSale = pcl.toLocaleString('vi-VN',{
-            style: 'currency',
-            currency: 'VND'
-        })
+        // const priceSale = pcl.toLocaleString('vi-VN',{
+        //     style: 'currency',
+        //     currency: 'VND'
+        // })
 
-        const price = pc.toLocaleString('vi-VN',{
-            style: 'currency',
-            currency: 'VND'
-        })
+        // const price = pc.toLocaleString('vi-VN',{
+        //     style: 'currency',
+        //     currency: 'VND'
+        // })
+
         return (
 
             <div className="productCard">
@@ -25,18 +29,29 @@ export default class ProductItem extends Component<Props, State> {
                 </div>
                 <div className="price">
                     <p>
-                        <span className="afterSale">{priceSale}</span>
-                        <span className="beforSale">{price}</span>
-
+                        <span className="afterSale">
+                            <NumberFormat value={this.props.product.salePrice} displayType={'text'} thousandSeparator={true} prefix={'VND'} />
+                        </span>
+                        <span className="beforSale">
+                            <NumberFormat value={this.props.product.price} displayType={'text'} thousandSeparator={true} prefix={'VND'} />
+                        </span>
                     </p>
                 </div>
-                <div className="button">
-                    <button className="btn btn-primary"
-                       /*  onClick={(event) => {
-                            cartService.addToCart(this.props.product.id)
-                            this.props.onNotifycation()
-                        }} */
-                    >Đưa vào giỏ</button>
+                <div className="bottom" style={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between' }}>
+
+                    <div className="detail">
+                        <Button>
+                            <SearchIcon />
+                        </Button>
+                    </div>
+                    <div className="button">
+                        <button className="btn btn-primary"
+                            // onClick={(event) => {
+                            //     cartService.addToCart(this.props.product.id)
+                            //     this.props.onNotifycation()
+                            // }}
+                        >Đưa vào giỏ</button>
+                    </div>
                 </div>
             </div>
         )
@@ -45,7 +60,7 @@ export default class ProductItem extends Component<Props, State> {
 
 export interface Props {
     product: Product;
-    onNotifycation() : void ;
+    // onNotifycation(): void;
 }
 
 export interface State {
